refactor: #13 StopTimerView와 SummaryView 분리
